1. Произвести автоматическое тестирование всех сервлетов.[#202349]

+++ chapter_009/src/main/java/ru/job4j/firsthttp/UserUpdateServlet.java
@@ -31,8 +31,8 @@ public class UserUpdateServlet extends HttpServlet {
      */
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        Upload upload = new Upload();
-        File repository = (File) this.getServletConfig().getServletContext().getAttribute("javax.servlet.context.tempdir");
+        Upload upload = Upload.getUploadInstance();
+        File repository = (File) this.getServletContext().getAttribute("javax.servlet.context.tempdir");
         Map<String, Object> fields = upload.getFields(req, repository);
         String name = (String) fields.get("name");
         String email = (String) fields.get("email");
@@ -46,11 +46,13 @@ public class UserUpdateServlet extends HttpServlet {
         String message;
         String photoPath = upload.uploadPhoto(photoId, repository);
         User savedUser = this.validate.findById(intOrDef(srgId, -1));
-        String savedPhoto = savedUser.getPhotoId();
-
+        String savedPhoto = "";
+        if (savedUser != null) {
+            savedPhoto = savedUser.getPhotoId();
+        }
         if (this.validate.update(new User(intOrDef(srgId, -1), name, email, photoPath, password, role))) {
             message = "The user was successfully updated.";
-            if (!photoPath.equals("")) {
+            if (!photoPath.equals("") && !savedPhoto.equals("")) {
                 new File(savedPhoto).delete();
             }
         } else {
@@ -63,8 +65,9 @@ public class UserUpdateServlet extends HttpServlet {
 
     /**
      * the method converts string to int or return def
+     *
      * @param srgId - the string for converting to int
-     * @param def - default int
+     * @param def   - default int
      * @return int
      */
     private int intOrDef(String srgId, int def) {
